+++ packages/build-plugin-alt/src/babelPluginMeta.ts
@@ -11,7 +11,7 @@ interface IMeta {
 interface IVisitorPass {
   opts?: {
     filename: string;
-    meta: IMeta;
+    meta?: IMeta;
   }
   filename?: string;
 }
@@ -40,10 +40,10 @@ const injectMetaVisitor: Visitor<IInjectVisitorPass> = {
         (BabelTypes.isStringLiteral(property) && property.value === 'pluginName')
       ) {
         const meta = pass.meta;
-        if (meta.pluginName) {
+        pass.cache.pluginNameAssignmentExists = true;
+        if (meta?.pluginName) {
           path.node.right = BabelTypes.stringLiteral(meta.pluginName);
           addInfoComment(path.node);
-          pass.cache.pluginNameAssignmentExists = true;
         }
       }
       if (
@@ -51,11 +51,11 @@ const injectMetaVisitor: Visitor<IInjectVisitorPass> = {
         (BabelTypes.isStringLiteral(property) && property.value === 'meta')
       ) {
         const meta = pass.meta;
-        if (meta.meta) {
+        pass.cache.metaAssignmentExists = true;
+        if (meta?.meta) {
           const generateAst = template(`const a = ${JSON.stringify(meta.meta)}`)() as BabelTypes.VariableDeclaration;
           path.node.right = generateAst.declarations[0].init;
           addInfoComment(path.node);
-          pass.cache.metaAssignmentExists = true;
         }
       }
     }
@@ -78,13 +78,13 @@ export default function (babel: { types: typeof BabelTypes }): {
         if (t.isIdentifier(path.node.declaration)) {
           const declarationName = path.node.declaration.name;
           path.parentPath.traverse(injectMetaVisitor, { ...state.opts, declarationName, cache })
-          if (!cache.pluginNameAssignmentExists) {
-            const generatedNode  = template(`${declarationName}.pluginName = "${state.opts.meta.pluginName}"`)();
+          if (!cache.pluginNameAssignmentExists && state.opts?.meta?.pluginName) {
+            const generatedNode  = template(`${declarationName}.pluginName = "${state.opts?.meta?.pluginName}"`)();
             addInfoComment(generatedNode);
             path.insertBefore(generatedNode);
           }
-          if (!cache.metaAssignmentExists) {
-            const generatedNode = template(`${declarationName}.meta = ${JSON.stringify(state.opts.meta.meta)}`)();
+          if (!cache.metaAssignmentExists && state.opts?.meta?.meta) {
+            const generatedNode = template(`${declarationName}.meta = ${JSON.stringify(state.opts?.meta?.meta)}`)();
             addInfoComment(generatedNode);
             path.insertBefore(generatedNode);
           }